Crate interlay_runtime_parachain

The interlay runtime. This can be compiled with #[no_std], ready for Wasm.

Re-exports

pub use btc_relay::bitcoin;

pub use primitives;

pub use primitives::CurrencyId::ForeignAsset;

pub use primitives::CurrencyId::LendToken;

pub use primitives::CurrencyId::Token;

pub use constants::currency::*;

pub use constants::time::*;

Modules

api

constants

A set of constant values used in the kintsugi runtime.

dex

evm

token_distribution

weights

xcm_config

Macros

construct_runtime

Construct a runtime, with the given name and the given pallets.

parameter_types

Create new implementations of the Get trait.

Structs

AdjustmentVariable

The adjustment variable of the runtime. Higher values will cause TargetBlockFullness to change the fees more rapidly.

AnnouncementDepositBase

AnnouncementDepositFactor

AssetAuthority

AssetMetadata

Data describing the asset properties.

BalanceWrapper

a wrapper around a balance, used in RPC to workaround a bug where using u128 in runtime-apis fails. See https://github.com/paritytech/substrate/issues/4641

BaseCallFilter

BasicDeposit

BlockHashCount

BlockNumberToBalance

CollatorPotId

CollatorSelectionAccount

CurrencyHooks

DealWithFees

DealWithRewards

DustRemovalWhitelist

EmissionPeriod

EnableManualSeal

EnactmentPeriod

EnsureKintsugiLabs

EscrowAnnuityAccount

EscrowAnnuityPalletId

EscrowBlockRewardProvider

ExecutiveBody

ExistentialDeposits

FarmingAccount

FarmingPalletId

FastTrackVotingPeriod

FeeAccount

FeePalletId

FieldDeposit

FixedU128

A fixed point number representation in the range. Fixed Point 128 bits unsigned, range = [0.000000000000000000, 340282366920938463463.374607431768211455]

GetDepositBase

GetDepositFactor

GetMaxSignatories

GetNativeCurrencyId

GetRelayChainCurrencyId

GetWrappedCurrencyId

IdentityFee

Implementor of WeightToFee that maps one unit of weight to one unit of fee.

InflationPeriod

IssueRequest

KintsugiLabsAccounts

LaunchPeriod

LoansAccount

LoansPalletId

MaxAdditionalFields

MaxAuthorities

MaxCandidates

MaxExpectedValue

MaxInvulnerables

MaxLocks

MaxNameLen

MaxPending

MaxPeriod

MaxProposalWeight

MaxProposals

MaxProxies

MaxRegistrars

MaxScheduledPerBlock

MaxSubAccounts

MaxVestingSchedules

MaxVotes

MaximumMultiplier

MaximumSchedulerWeight

MinCandidates

MinVestedTransfer

MinimumDeposit

MinimumMultiplier

Minimum amount of the multiplier. This value cannot be too low. A test case should ensure that combined with AdjustmentVariable, we can recover from the minimum. See multiplier_can_grow_from_zero.

MinimumPeriod

Offset

OperationalFeeMultiplier

This value increases the priority of Operational transactions by adding a “virtual tip” that’s equal to the OperationalFeeMultiplier * final_fee.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParachainBlocksPerBitcoinBlock

PauseTooLongNames

Perbill

A fixed point representation of a number in the range [0, 1].

Period

Permill

A fixed point representation of a number in the range [0, 1].

PreimageBaseDepositz

PreimageByteDepositz

ProxyDepositBase

ProxyDepositFactor

RedeemRequest

ReplaceRequest

ReservedDmpWeight

ReservedXcmpWeight

RewardPeriod

RocksDbWeight

By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SessionKeys

SessionLength

Span

SubAccountDeposit

SupplyAccount

SupplyPalletId

TargetBlockFullness

The portion of the NORMAL_DISPATCH_RATIO that we adjust the fees with. Blocks filled less than this will decrease the weight and more will increase.

TechnicalCommitteeMaxMembers

TechnicalCommitteeMaxProposals

TechnicalCommitteeMotionDuration

TotalWrapped

TransactionByteFee

TransactionConverter

TreasuryAccount

TreasuryPalletId

VaultAnnuityAccount

VaultAnnuityPalletId

VaultBlockRewardProvider

VaultRegistryAccount

VaultRegistryPalletId

Version

VotingPeriod

Weight

Enums

BtcRelayCall

Contains a variant per dispatchable extrinsic that this pallet has.

DispatchClass

A generalized group of dispatch types.

Everything

A [Contains] implementation that contains every value.

NominationEvent

The Event enum of this pallet

Nothing

A [Contains] implementation that contains no value.

OriginCaller

ProxyType

The type used to represent the kinds of proxying allowed.

RuntimeCall

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

TimestampCall

Contains a variant per dispatchable extrinsic that this pallet has.

Constants

MAXIMUM_BLOCK_WEIGHT

We allow for 0.5 seconds of compute with a 12 second average block time.

TARGET_SPACING

Target Spacing: 10 minutes (600 seconds)

VERSION

This runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

WEIGHT_REF_TIME_PER_SECOND

Traits

BuildStorage

Complex storage builder stuff.

CurrencyInfo

Get

A trait for querying a single value from a type.

KeyOwnerProofSystem

Something which can compute and check proofs of a historical key owner and return full identification data of that key owner.

StorageValue

A trait for working with macro-generated storage values under the substrate storage API.

Functions

get_all_module_accounts

native_version

The version information used to identify this runtime when compiled natively.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

AssetRegistry

AssetRegistryConfig

Aura

AuraConfig

AuraExt

AuraExtConfig

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Authorship

BTCRelay

BTCRelayConfig

Balance

Balance of an account.

BaseFee

BaseFeeConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

CheckedExtrinsic

Extrinsic type that has already been checked.

ClientsInfo

CollatorSelection

CollatorSelectionConfig

CollatorSelectionUpdateOrigin

We allow root and the Relay Chain council to execute privileged collator selection operations.

CumulusXcm

Currency

Democracy

DemocracyConfig

DexGeneral

DexStable

DexSwapRouter

DmpQueue

EVM

EVMChainId

EVMChainIdConfig

EVMConfig

Escrow

EscrowAnnuity

EscrowAnnuityInstance

EscrowRewards

EscrowRewardsInstance

Ethereum

EthereumConfig

Executive

Executive: handles dispatch to the various modules.

Farming

FarmingRewards

Fee

FeeConfig

GenesisConfigDeprecated

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

Identity

Issue

IssueConfig

Liquidity

Loans

LoansConfig

LockIdentifier

An identifier for a lock. Used for disambiguating different locks so that they can be individually replaced or removed.

Moment

An instant or duration in time.

MultiTransactionPayment

Multisig

Nomination

NominationConfig

Nonce

Index of a transaction in the chain. 32-bit should be plenty.

Oracle

OracleConfig

OracleName

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PolkadotXcmConfig

Preimage

Proxy

Rate

Ratio

Redeem

RedeemConfig

Replace

ReplaceConfig

Scheduler

Security

Session

SessionConfig

Shortfall

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedFixedPoint

The signed fixed point type.

SignedInner

The Inner type of the SignedFixedPoint.

SlowAdjustingFeeUpdate

StablePoolId

Sudo

SudoConfig

Supply

SupplyConfig

System

SystemConfig

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalCommitteeInstance

TechnicalMembership

TechnicalMembershipConfig

Timestamp

Tokens

TokensConfig

TransactionPayment

TxPause

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

UnknownTokens

UnsignedFixedPoint

The unsigned fixed point type.

UnsignedInner

The Inner type of the UnsignedFixedPoint.

Utility

VaultAnnuity

VaultAnnuityInstance

VaultCapacity

VaultCapacityInstance

VaultRegistry

VaultRegistryConfig

VaultRewards

VaultRewardsInstance

VaultStaking

Vesting

VestingConfig

XTokens

XcmpQueue