Module interlay_runtime_parachain::xcm_config

Re-exports

pub use xcm_executor;

Structs

AccountIdToMultiLocation

Ancestry

CanonicalizedIbtcPerSecond

CanonicalizedIntrPerSecond

CurrencyIdConvert

DotPerSecond

IbtcPerSecond

IntrPerSecond

MaxAssetsForTransfer

MaxAssetsIntoHolding

MaxInstructions

MyFixedConversionRateProvider

ParachainMinFee

ParentLocation

ParentNetwork

RelayChainOrigin

RelayNetwork

SafeCallFilter

A call filter for the XCM Transact instruction. This is a temporary measure until we properly account for proof size weights.

SelfLocation

ToAuthor

UnitWeightCost

UniversalLocation

XcmConfig

Functions

canonical_currency_location

dot_per_second

non_canonical_currency_location

Type Definitions

Barrier

LocalAssetTransactor

LocalOriginToLocation

No local origins on this chain are allowed to dispatch XCM sends/executions.

Trader

XcmOriginToTransactDispatchOrigin

This is the type we use to convert an (incoming) XCM origin into a local Origin instance, ready for dispatching a transaction with Xcm’s Transact. There is an OriginKind which can biases the kind of local Origin it will become.

XcmRouter

The means for routing XCM messages which are not for local execution into the right message queues.